Separated the collision from the visuals
Now you can create layers (which are textures) to draw on.

Also the tilemaps no longer have renderers are only for collision.

The visuals are done by the textures so we can have players drawing in the background or foreground not affecting the collisions.

Also because the visuals are on a texture now, we can apply antialiasing or shaders to modify the look which would have been significantly harder or impossible with the tilemap solution.